allign data

+++ include/CreateWithParameters.h
@@ -6,28 +6,28 @@ namespace OperationArchitecture
     class ICreateWithParameters
     {
         public:
-        virtual IOperationBase* Create(const void *config, unsigned int &sizeOut) = 0;
+        virtual IOperationBase* Create(const void *config, size_t &sizeOut) = 0;
     };
 
     template<typename... PARAMS>
     class CreateWithParameters : public ICreateWithParameters
     {
         protected:
-        IOperationBase*(*_factory)(const void *, unsigned int &, PARAMS...);
+        IOperationBase*(*_factory)(const void *, size_t &, PARAMS...);
 		std::tuple<PARAMS...> _params;
 
 		template<std::size_t... Is>
-		IOperationBase* CreateWithTuple(const void *config, unsigned int &sizeOut, const std::tuple<PARAMS...>& tuple, std::index_sequence<Is...>) {
+		IOperationBase* CreateWithTuple(const void *config, size_t &sizeOut, const std::tuple<PARAMS...>& tuple, std::index_sequence<Is...>) {
 			return _factory(config, sizeOut, std::get<Is>(tuple)...);
 		}
 
         public:
-		CreateWithParameters(IOperationBase*(*factory)(const void *, unsigned int &, PARAMS...), PARAMS... params)
+		CreateWithParameters(IOperationBase*(*factory)(const void *, size_t &, PARAMS...), PARAMS... params)
 		{
             _factory = factory;
 			_params = std::tuple<PARAMS...>(params...);
 		}
-        IOperationBase* Create(const void *config, unsigned int &sizeOut) override
+        IOperationBase* Create(const void *config, size_t &sizeOut) override
         {
             return CreateWithTuple(config, sizeOut, _params, std::index_sequence_for<PARAMS...>());
         }
