+++ src/Operations/Operation_Polynomial.cpp
@@ -25,7 +25,7 @@ namespace OperationArchitecture
 		variables[0]->Set(val);
 	}
 
-	IOperationBase *Operation_Polynomial::Create(const void *config, unsigned int &sizeOut)
+	IOperationBase *Operation_Polynomial::Create(const void *config, size_t &sizeOut)
 	{
 		const Operation_PolynomialConfig *polynomialConfig = Config::CastConfigAndOffset<Operation_PolynomialConfig>(config, sizeOut);
 		return new Operation_Polynomial(polynomialConfig);
